fiz novos desafios

+++ desafios/d003/continha.py
@@ -1,9 +1,8 @@
 print('====== DESAFIO 03 ======')
 
-num1 = float(input('Informe um número: '))
-num2 = float(input('Informe outro número: '))
-
+num1 = int(input('Informe um número: '))
+num2 = int(input('Informe outro número: '))
 soma = num1 + num2
-
 print('O resultado da soma é: ', soma)
+
